Introduce absorbable impl for iterable types
