+++ src/poseidon/mod.rs
@@ -279,47 +279,3 @@ impl<F: PrimeField> CryptographicSponge<F> for PoseidonSponge<F> {
         squeezed_elems
     }
 }
-
-/*
-#[cfg(test)]
-pub mod tests {
-    use crate::poseidon::PoseidonSponge;
-    use crate::CryptographicSponge;
-    use ark_ed_on_bls12_381::{Fq, Fr};
-    use ark_ff::{to_bytes, One};
-    use ark_marlin::fiat_shamir::poseidon::PoseidonSponge as FSPoseidonSponge;
-    use ark_marlin::fiat_shamir::{FiatShamirAlgebraicSpongeRng, FiatShamirRng};
-    use crate::poseidon::constraints::PoseidonSpongeVar;
-    use ark_relations::r1cs::ConstraintSystem;
-    use crate::constraints::CryptographicSpongeVar;
-    use ark_r1cs_std::fields::fp::FpVar;
-    use ark_r1cs_std::fields::FieldVar;
-    use ark_r1cs_std::R1CSVar;
-
-    type F = Fr;
-    type CF = Fq;
-
-    #[test]
-    fn test() {
-        let mut s = PoseidonSponge::<CF>::new();
-        s.absorb(&CF::one());
-        let output = s.squeeze_bytes(180);
-        println!("{:?}", output);
-
-        /*
-        let mut s = FiatShamirAlgebraicSpongeRng::<F, CF, FSPoseidonSponge<CF>>::new();
-        s.absorb_native_field_elements(&[CF::one()]);
-        let output = s.squeeze_native_field_elements(1);
-        println!("{:?}", output);
-
-         */
-
-        let cs = ConstraintSystem::<CF>::new_ref();
-        let mut s = PoseidonSpongeVar::new(cs);
-        s.absorb(&[FpVar::<CF>::one()]).unwrap();
-        let output = s.squeeze_byte_vars(180).unwrap();
-        println!("{:?}", output.value().unwrap());
-    }
-}
-
- */
